`backends.tests.FkConstraintsTests.test_check_constraints`:
Database | Before | After
--- | --- | ---
MySQL | 720s | 0.78 s
MariaDB | 188s | 0.65 s