@carltongibson 2021-03-24
So was it quicker? e.g. https://djangoci.com/job/pull-requests-bionic/database=mysql,label=bionic-pr,python=python3.8/10902/ still took 20mins? 🤔