@felixxm 2021-03-24
I compared times of a single test `backends.tests.FkConstraintsTests.test_check_constraints` :smile: